+++ src/main/java/org/tron/common/logsfilter/capsule/ContractEventTriggerCapsule.java
@@ -3,7 +3,7 @@ package org.tron.common.logsfilter.capsule;
 import java.util.List;
 import lombok.Getter;
 import lombok.Setter;
-import org.tron.common.logsfilter.ContractEventParser;
+import org.tron.common.logsfilter.ContractEventParserAbi;
 import org.tron.common.logsfilter.EventPluginLoader;
 import org.tron.common.logsfilter.FilterQuery;
 import org.tron.common.logsfilter.trigger.ContractEventTrigger;
@@ -54,8 +54,9 @@ public class ContractEventTriggerCapsule extends TriggerCapsule {
 
   @Override
   public void processTrigger() {
-    contractEventTrigger.setTopicMap(ContractEventParser.parseTopics(topicList, abiEntry));
-    contractEventTrigger.setDataMap(ContractEventParser.parseEventData(data, topicList, abiEntry));
+    contractEventTrigger.setTopicMap(ContractEventParserAbi.parseTopics(topicList, abiEntry));
+    contractEventTrigger
+        .setDataMap(ContractEventParserAbi.parseEventData(data, topicList, abiEntry));
 
     if (FilterQuery.matchFilter(contractEventTrigger)) {
       EventPluginLoader.getInstance().postContractEventTrigger(contractEventTrigger);
